Remove useless tuples which have only null as attributes
In the case of drawing a line chart for materials from log files of "Sandanshiki Kanpan", the data may cause line disconnection, because X-axis (date and time) may have null as Y-axis (quantity of materials).
It is really hard for us to stand all line disconnection.
Therefore, projected relation removes a tuple which have only null as attributes.